Reduce repeated parsing attempts on tasks [RHELDST-12184]
Previously, the parser was called to parse arguments too frequently
(when args or extra_args were called or anytime after if they held empty
structures). This commit implements a few changes to reduce the number
of parser calls and maintain the integrity of the original argument list
should multiple calls be unavoidable.