Resolve python2.7-related exodus-rsync log buffering [RHELDST-14217]
Prior to this commit, the pubtools-exodus-push CLI would buffer the
exodus-rsync output. This buffering issue was only present when
invoking pubtools-exodus-push with Python 2.7.

The exodus-rsync output is now logged in a way that is compatible
with both Python 2.7 and Python 3.X. Buffering is no longer present,
regardless of the Python version used to invoke pubtools-exodus-push.