Use --exodus-commit=phase1 in pulp hook [RHELDST-20490]
When this hook is active, the control flow is:

1. pubtools-exodus: create publish
2. rhsm-pulp: add content to publish
3. pubtools-exodus: commit publish

Problem: step 2 might succeed from Pulp's point of view, but then fail
at step 3 leaving the content not actually published to the CDN storage.
As Pulp believes the publishes were successful, it would then skip
publish of certain files at the next attempt, leading to missing
content.

The newly introduced concept of a phase1 commit should fix this by
allowing step 2 to request that all "phase 1 content" (e.g. RPMs, but
not repodata) are flushed to the CDN storage before proceeding. Go ahead
and start using that via the new related exodus-rsync argument.

While it would appear to make sense to do this at *every* publish, an
env var is added as an escape hatch to go back to the old behavior, just
in case something doesn't work as expected or upgrades are performed in
the wrong order.