Drop mypy from "tox -e static"
mypy started to complain with:

    pubtools/exodus/gateway.py:8: error: Library stubs not installed for "requests.packages.urllib3.util.retry"  [import-untyped]
    pubtools/exodus/gateway.py:8: note: Hint: "python3 -m pip install types-requests"

...despite that types-requests *is* already installed. Other projects
have seen similar issues.

As we do not seem to detect any genuine issues with mypy in practice,
let's drop it from the default CI configuration to reduce the CI
maintenance cost. mypy currently remains in the test-requirements.txt
file to enable running the tool locally.