Fix certain requests not retrying
Per [1], the Retry class defaults to allowing retries only on certain
methods, not including POST.

That makes sense since POST requests are not generally expected to be
idempotent. However, we've intentionally designed the relevant exodus-gw
APIs (create/commit publish) to be retry-safe anyway. Enable retries of
POST so that temporary failures on those endpoints can also be recovered
from.

[1] https://urllib3.readthedocs.io/en/stable/reference/urllib3.util.html#urllib3.util.Retry